Change bash prompt to include server name
I've changed the bash prompt to include the $SERVER_NAME variable, which will be the canonical domain name of the WordPress installation. This replaces the default prompt which uses the instance's hostname – which in this case is an auto-assigned name from docker, and doesn't map to an instance's URL or environment.

This will make it clearer to see which website and environment you're using the shell for. Without this, the auto-assigned docker names could be confused between environments, resulting in scripts being run in the wrong environment or against the wrong WordPress website.