Generate the Pingdom IP address whitelist dynamically at build time
This commit replaces the hardcoded Pingdom whitelist file with a script which will automatically generate the whitelist at image build time.

The hardcoded whitelist has been problematic because Pingdom frequently change their IP ranges, and manual intervention was required to update the file and commit it into this repository. By offloading this responsibility to the new build-time script, we benefit from reduced maintenance overhead as we can simply rebuild the docker image to update the Pingdom IP address whitelist. No editing of files or commits to git necessary.