Allow ImageMagick to read PDF files
This commit adds a `policy.xml` configuration file for ImageMagick, which will allow it to read and process PDF files.

The default ImageMagick configuration has recently changed to stop ImageMagick from reading PDF files. Previously, the default config *did* allow it, which meant there was no need for us to override it here.